backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
prometheus-client==0.21.0
psutil==6.1.0
